shared: move lookup utils from kubevirt-plugin
