+++ frontend/packages/console-shared/src/index.ts
@@ -2,3 +2,4 @@ export * from './components';
 export * from './constants';
 export * from './selectors';
 export * from './types';
+export * from './utils';
