+++ frontend/packages/console-shared/src/utils/utils.ts
@@ -1,0 +1,28 @@
+import { K8sResourceKind } from '@console/internal/module/k8s';
+import { FirehoseResult } from '@console/internal/components/utils';
+import { getUID } from '../selectors';
+
+export type EntityMap<A> = { [propertyName: string]: A };
+export type K8sEntityMap<A extends K8sResourceKind> = EntityMap<A>;
+
+type KeyResolver<A> = (entity: A) => string;
+
+export const createBasicLookup = <A>(list: A[], getKey: KeyResolver<A>): EntityMap<A> => {
+  return (list || []).reduce((lookup, entity) => {
+    const key = getKey(entity);
+    if (key) {
+      lookup[key] = entity;
+    }
+    return lookup;
+  }, {});
+};
+
+export const createLookup = <A extends K8sResourceKind>(
+  loadingList: FirehoseResult<A[]>,
+  getKey?: KeyResolver<A>,
+): K8sEntityMap<A> => {
+  if (loadingList && loadingList.loaded) {
+    return createBasicLookup(loadingList.data, getKey || getUID);
+  }
+  return {};
+};
