+++ frontend/packages/kubevirt-plugin/src/k8s/requests/vmi/migration.ts
@@ -3,9 +3,9 @@ import { k8sCreate } from '@console/internal/module/k8s';
 import { VMIKind } from '../../../types/vm';
 import { VirtualMachineInstanceMigrationModel } from '../../../models';
 import { Migration } from './objects/migration';
-import { prefixedId } from '../../../utils';
+import { prefixedID } from '../../../utils';
 
-export const getMigrationName = (vmi: VMIKind) => prefixedId(getName(vmi), 'migration');
+export const getMigrationName = (vmi: VMIKind) => prefixedID(getName(vmi), 'migration');
 
 export const startVMIMigration = (vmi: VMIKind) => {
   const migration = new Migration().setName(getMigrationName(vmi)).setVMI(vmi);
