+++ frontend/packages/kubevirt-plugin/src/selectors/vm/combined.ts
@@ -4,7 +4,7 @@ import {
   VM_STATUS_V2V_CONVERSION_IN_PROGRESS,
 } from 'kubevirt-web-ui-components';
 
-import { getName } from '@console/shared';
+import { getName, createBasicLookup } from '@console/shared';
 import { K8sResourceKind } from '@console/internal/module/k8s';
 
 import { VMIKind, VMKind } from '../../types/vm';
@@ -12,7 +12,6 @@ import { getUsedNetworks, isVMRunning } from './selectors';
 import { VMMultiStatus } from '../../types';
 import { Network } from './types';
 import { NetworkType, POD_NETWORK } from '../../constants/vm';
-import { createBasicLookup } from '../../utils';
 
 const IMPORTING_STATUSES = new Set([VM_STATUS_IMPORTING, VM_STATUS_V2V_CONVERSION_IN_PROGRESS]);
