+++ frontend/packages/kubevirt-plugin/src/types/types.ts
@@ -1,10 +1,7 @@
-import { K8sResourceKind, PodKind, TemplateKind } from '@console/internal/module/k8s';
+import { PodKind, TemplateKind } from '@console/internal/module/k8s';
 import { VMKind } from './vm';
 
 export type VMLikeEntityKind = VMKind | TemplateKind;
-
-export type EntityMap<A> = { [propertyName: string]: A };
-export type K8sEntityMap<A extends K8sResourceKind> = EntityMap<A>;
 
 export type VMMultiStatus = {
   status: string;
